FIX no redundant Class Limiter
