+++ honeygrove/services/FTPService.py
@@ -6,14 +6,13 @@ from twisted.cred.portal import Portal
 from twisted.protocols.ftp import *
 from twisted.protocols.ftp import _FileReader as FR
 from twisted.protocols.ftp import _FileWriter as FW
-from twisted.protocols.policies import WrappingFactory
 
 from honeygrove import config
 from honeygrove.core.FilesystemParser import FilesystemParser
 from honeygrove.core.HoneytokenDB import HoneytokenDataBase
 from honeygrove.logging import log as logging
 from honeygrove.services.ServiceBaseModel import ServiceBaseModel
-
+from honeygrove.services.ServiceBaseModel import Limiter
 
 class FTPService(ServiceBaseModel):
 
@@ -30,13 +29,13 @@ class FTPService(ServiceBaseModel):
 
         self._fService = FTPFactory(portal)
 
-        self._limiter = Limiter(self._fService)
+        self._name = config.ftpName
+        self._port = config.ftpPort
+
+        self._limiter = Limiter(self._fService, self._name, config.FTP_conn_per_host)
 
         self.protocol = FTPProtocol
         self._fService.protocol = self.protocol
-
-        self._name = config.ftpName
-        self._port = config.ftpPort
 
     def startService(self):
         self._stop = False
@@ -46,27 +45,6 @@ class FTPService(ServiceBaseModel):
         self._stop = True
         self._transport.stopListening()
 
-class Limiter(WrappingFactory):
-
-    maxConnectionsPerPeer = config.FTP_conn_per_host
-
-    def startFactory(self):
-        self.peerConnections = {}
-
-    def buildProtocol(self, addr):
-        peerHost = addr.host
-        connectionsCount = self.peerConnections.get(peerHost, 0)
-        if connectionsCount >= self.maxConnectionsPerPeer:
-            logging.limit_reached(config.ftpName, peerHost)
-            return None
-        self.peerConnections[peerHost] = connectionsCount + 1
-        return WrappingFactory.buildProtocol(self, addr)
-
-    def unregisterProtocol(self, p):
-        peerHost = p.getPeer().host
-        self.peerConnections[peerHost] -= 1
-        if self.peerConnections[peerHost] == 0:
-            del self.peerConnections[peerHost]
 
 class FTPProtocol(FTP):
     overwritten_commands_whitelist = ['CWD', 'DELE', 'LIST', 'MDTM', 'MKD',
